Fix flake8 incompatibilities
