+++ braindecode/augmentation/transforms.py
@@ -745,7 +745,7 @@ class RandomSensorsRotation(Transform):
                 sensors_positions_matrix
             )
         assert isinstance(sensors_positions_matrix, torch.Tensor),\
-           "sensors_positions should be an Tensor"
+            "sensors_positions should be an Tensor"
         assert isinstance(max_degrees, Real) and max_degrees >= 0,\
             "max_degrees should be non-negative float."
         assert isinstance(axis, str) and axis in ['x', 'y', 'z'],\
